Refactor : 패키지 정리, 불필요한 파일 삭제 #2
