+++ src/main/java/com/sendback/domain/auth/dto/request/AuthRequestDto.java
@@ -1,5 +1,7 @@
 package com.sendback.domain.auth.dto.request;
 
+import jakarta.validation.constraints.NotBlank;
+
 public class AuthRequestDto{
     public record RefreshTokenRequestDto(String refreshToken){}
