+++ src/main/java/com/sendback/domain/auth/service/AuthService.java
@@ -3,7 +3,7 @@ package com.sendback.domain.auth.service;
 import com.sendback.domain.auth.dto.Token;
 import com.sendback.global.config.redis.RedisService;
 import com.sendback.global.config.jwt.JwtProvider;
-import com.sendback.global.error.type.UnAuthorizedException;
+import com.sendback.global.exception.type.UnAuthorizedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -21,6 +21,7 @@ public class AuthService {
     @Transactional(noRollbackFor = UnAuthorizedException.class)
     public Token reissueToken(String refreshToken) {
         Long userId = jwtProvider.parseRefreshToken(refreshToken);
+        validateRefreshToken(refreshToken, userId);
         Token tokens = jwtProvider.issueToken(userId);
         redisService.put(userId, tokens.refreshToken(), REFRESH_TOKEN_EXPIRE_TIME);
         return tokens;
@@ -28,5 +29,15 @@ public class AuthService {
 
     public void logoutSocial(Long userId){
         redisService.delete(userId);
+    }
+
+    private void validateRefreshToken(String refreshToken, Long userId) {
+        try {
+            jwtProvider.validateRefreshToken(refreshToken);
+            jwtProvider.equalsRefreshToken(userId, refreshToken);
+        } catch (UnAuthorizedException e) {
+            redisService.delete(userId);
+            throw e;
+        }
     }
 }
