--- src/main/java/com/sendback/domain/auth/type/UserType.java
@@ -1,10 +1,0 @@
-package com.sendback.domain.auth.type;
-
-public enum UserType {
-    KAKAO,
-    NAVER,
-    PAYCO,
-    GOOGLE,
-    FACEBOOK,
-    NORMAL
-}
