--- src/main/java/com/sendback/global/error/type/ForbiddenException.java
@@ -1,12 +1,0 @@
-package com.sendback.global.error.type;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.FORBIDDEN)
-public class ForbiddenException extends RuntimeException {
-
-	public ForbiddenException(String message) {
-		super(message);
-	}
-}
