--- src/main/java/com/sendback/global/error/type/InternalServerErrorException.java
@@ -1,12 +1,0 @@
-package com.sendback.global.error.type;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
-public class InternalServerErrorException extends RuntimeException {
-
-	public InternalServerErrorException(String message) {
-		super(message);
-	}
-}
