--- src/main/java/com/sendback/global/error/type/NotFoundException.java
@@ -1,12 +1,0 @@
-package com.sendback.global.error.type;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class NotFoundException extends RuntimeException {
-
-	public NotFoundException(String message) {
-		super(message);
-	}
-}
